improve RefBuilder
