+++ celements-model/src/main/java/com/celements/model/reference/RefBuilder.java
@@ -1,0 +1,140 @@
+package com.celements.model.reference;
+
+import static com.celements.model.util.EntityTypeUtil.*;
+import static com.celements.model.util.References.*;
+import static com.google.common.base.Preconditions.*;
+import static com.google.common.base.Strings.*;
+
+import java.util.TreeMap;
+
+import org.xwiki.model.EntityType;
+import org.xwiki.model.reference.EntityReference;
+
+public class RefBuilder {
+
+  private final TreeMap<EntityType, EntityReference> refs;
+  private boolean nullable;
+
+  public RefBuilder() {
+    refs = new TreeMap<>();
+    nullable = false;
+  }
+
+  public int depth() {
+    return refs.size();
+  }
+
+  public RefBuilder nullable() {
+    this.nullable = true;
+    return this;
+  }
+
+  public RefBuilder wiki(String name) {
+    return with(EntityType.WIKI, name);
+  }
+
+  public RefBuilder space(String name) {
+    return with(EntityType.SPACE, name);
+  }
+
+  public RefBuilder doc(String name) {
+    return with(EntityType.DOCUMENT, name);
+  }
+
+  public RefBuilder att(String name) {
+    return with(EntityType.ATTACHMENT, name);
+  }
+
+  public RefBuilder with(EntityReference ref) {
+    while (ref != null) {
+      with(ref.getType(), ref.getName());
+      ref = ref.getParent();
+    }
+    return this;
+  }
+
+  public RefBuilder with(EntityType type, String name) {
+    if (type != null) {
+      if (!isNullOrEmpty(name)) {
+        refs.put(type, new EntityReference(name, type));
+      } else {
+        refs.remove(type);
+      }
+    }
+    return this;
+  }
+
+  /**
+   * @return absolute reference, null if insufficient values and {@link #nullable()}
+   * @throws IllegalArgumentException
+   *           if insufficient values and not {@link #nullable()}
+   */
+  public EntityReference build() {
+    return build(!refs.isEmpty() ? refs.descendingMap().firstKey() : null);
+  }
+
+  /**
+   * @param type
+   * @return absolute reference of given class, null if insufficient values and {@link #nullable()}
+   * @throws IllegalArgumentException
+   *           if insufficient values and not {@link #nullable()}
+   */
+  public EntityReference build(EntityType type) {
+    return build((type != null) ? getClassForEntityType(type) : EntityReference.class);
+  }
+
+  /**
+   * @param token
+   * @return absolute reference of given class, null if insufficient values and {@link #nullable()}
+   * @throws IllegalArgumentException
+   *           if insufficient values and not {@link #nullable()}
+   */
+  public <T extends EntityReference> T build(Class<T> token) {
+    T ref = null;
+    try {
+      EntityReference relativeRef = buildRelative(getEntityTypeForClass(token).orNull());
+      if (relativeRef != null) {
+        // clone for absolute, immutability reference
+        ref = cloneRef(relativeRef, token);
+      }
+    } catch (IllegalArgumentException iae) {
+      if (!nullable) {
+        throw iae;
+      }
+    }
+    return ref;
+  }
+
+  /**
+   * @return relative reference, null if insufficient values and {@link #nullable()}
+   * @throws IllegalArgumentException
+   *           if insufficient values and not {@link #nullable()}
+   */
+  public EntityReference buildRelative() {
+    return buildRelative(null);
+  }
+
+  /**
+   * @param type
+   * @return relative reference from type, null if insufficient values and {@link #nullable()}
+   * @throws IllegalArgumentException
+   *           if insufficient values and not {@link #nullable()}
+   */
+  public EntityReference buildRelative(EntityType type) {
+    EntityReference ret = null;
+    EntityReference lastRef = null;
+    for (EntityReference ref : refs.descendingMap().values()) {
+      if ((type == null) || (type.compareTo(ref.getType()) >= 0)) {
+        if (ret == null) {
+          ret = ref;
+        } else {
+          lastRef.setParent(ref);
+        }
+        lastRef = ref;
+      }
+    }
+    checkArgument(nullable || (ret != null), "missing information for building reference");
+    return ret;
+  }
+
+}
